+++ src/commands.js
@@ -13,6 +13,7 @@ module.exports = [
   { chained: true, name: 'restart', description: 'Restart your own application' },
   { chained: true, name: 'show', description: 'Displays active platform' },
   { chained: false, name: 'docbuild', description: 'Build docs' },
+  { chained: false, name: 'format', description: 'Format a Freyfile (overwriting existing, so only do this under source control)' },
   { chained: false, name: 'restore', description: 'Restore latest state backup' },
   { chained: false, name: 'destroy', description: 'Destroy all that is in state' },
   { chained: false, name: 'remote', description: 'Execute a remote command - or opens console' }