+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Unreleased and unplanned todos
 - [ ] Convert `format.js` to ES6 and make it a first class citizen via `frey format`
 - [ ] Check for `git ignore Frey-residu-*`
 - [ ] Offer cast5 encryption and decryption for `env.sh` (then remove `rebuild-env.sh` script from dependents)
+- [x] Convert `format.js` to ES6 and make it a first class citizen via `frey format`
 - [x] Implement Backup and Restore via Ansible
 - [x] Upgrade to Ansible 2.0.1.0 as that fixes an issue with sync asking for sudo password https://github.com/ansible/ansible-modules-core/issues/2156
 - [x] Remove invalid underscore prefix from a few public `Shell` methods